# api/swagger.yml
openapi: "3.0.3"
info:
  title: Messanger api
  version: "1.0"
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /chats:
    get:
      summary: Get chats infos for user
      operationId: listChats
      x-ogen-operation-group: Chats
      tags:
        - chats
      responses:
        "200":
          $ref: "#/components/responses/chatsResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
    post:
      summary: Create new chat
      operationId: createChat
      x-ogen-operation-group: Chats
      tags:
        - chats
      requestBody:
        $ref: "#/components/requestBodies/chatBody"
      responses:
        "200":
          $ref: "#/components/responses/chatResponse"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/join:
    post:
      summary: Join chat
      operationId: joinChat
      x-ogen-operation-group: Chats
      tags:
        - chats
      requestBody:
        $ref: "#/components/requestBodies/joinChatBody"
      responses:
        "204":
          description: "you joined chat"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "409":
          description: you have already joined this chat
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/{chatId}:
    get:
      summary: Get chat info by id
      operationId: getChatById
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
      responses:
        "200":
          $ref: "#/components/responses/chatResponse"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
    put:
      summary: Update chat info
      operationId: updateChat
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
      requestBody:
        $ref: "#/components/requestBodies/chatBody"
      responses:
        "200":
          $ref: "#/components/responses/chatResponse"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
    delete:
      summary: Delete chat by id
      operationId: deleteChat
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
      responses:
        "204":
          description: "chat deleted"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/{chatId}/leave:
    post:
      summary: Leave chat by id
      operationId: leaveChat
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
      responses:
        "204":
          description: "you left chat"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/{chatId}/join-code:
    get:
      summary: Get join code for chat by id
      operationId: getJoinCode
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
      responses:
        "200":
          $ref: "#/components/responses/joinCodeResponse"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/{chatId}/members:
    get:
      summary: get members for chat
      operationId: listMembers
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
      responses:
        "200":
          $ref: "#/components/responses/membersResponse"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/{chatId}/members/banned:
    get:
      summary: get banned members for chat
      operationId: listBannedUsers
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
      responses:
        "200":
          $ref: "#/components/responses/bannedMembersResponse"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/{chatId}/members/{userId}/ban:
    post:
      summary: ban user in chat
      operationId: banUser
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
        - $ref: "#/components/parameters/userId"
      responses:
        "204":
          description: user banned
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          description: chat or member not found
        "409":
          description: user has already been banned
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/{chatId}/members/{userId}/unban:
    post:
      summary: Unban user in chat
      operationId: UnbanUser
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
        - $ref: "#/components/parameters/userId"
      responses:
        "204":
          description: user banned
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          description: chat or member not found
        "409":
          description: user is not banned
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /chats/{chatId}/members/{userId}/set-role:
    post:
      summary: Set role to user
      operationId: setRole
      x-ogen-operation-group: Chats
      tags:
        - chats
      parameters:
        - $ref: "#/components/parameters/chatId"
        - $ref: "#/components/parameters/userId"
      requestBody:
        $ref: "#/components/requestBodies/setRoleBody"
      responses:
        "204":
          description: role setted
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          description: chat or member or role not found
        "500":
          $ref: "#/components/responses/internalErrorResponse"
  /roles:
    get:
      summary: get roles for chat
      operationId: listRoles
      x-ogen-operation-group: Roles
      tags:
        - roles
      parameters:
        - $ref: "#/components/parameters/chatIdInQuery"
      responses:
        "200":
          $ref: "#/components/responses/rolesResponse"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "500":
          $ref: "#/components/responses/internalErrorResponse"
    post:
      summary: Create role in Chat
      operationId: createRole
      x-ogen-operation-group: Roles
      tags:
        - roles
      parameters:
        - $ref: "#/components/parameters/chatIdInQuery"
      requestBody:
        $ref: "#/components/requestBodies/roleBody"
      responses:
        "200":
          $ref: "#/components/responses/roleResponse"
        "400":
          $ref: "#/components/responses/invalidInputResponse"
        "401":
          $ref: "#/components/responses/unauthenticatedResponse"
        "403":
          $ref: "#/components/responses/unauthorizedResponse"
        "404":
          $ref: "#/components/responses/chatNotFoundResponse"
        "409":
          description: role with this name already exists
        "500":
          $ref: "#/components/responses/internalErrorResponse"
...
components:
  responses:
    invalidInputResponse:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    unauthenticatedResponse:
      description: Unauthenticated
    unauthorizedResponse:
      description: Access forbidden
    internalErrorResponse:
      description: Internal server error
...
    chatNotFoundResponse:
      description: Chat not found
...
    chatsResponse:
      description: Chats
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "./schemas.yml#/components/schemas/chat"
    chatResponse:
      description: chat info response
      content:
        application/json:
          schema:
            $ref: "./schemas.yml#/components/schemas/chat"
    joinCodeResponse:
      description: get join code response
      content:
        application/json:
          schema:
            type: object
            properties:
              join_code:
                type: string
            required:
              - join_code
    membersResponse:
      description: get members response
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "./schemas.yml#/components/schemas/member"
    bannedMembersResponse:
      description: banned members response
      content:
        application/json:
          schema:
            type: array
            items:
              type: object
              properties:
                user_id:
                  $ref: "./schemas.yml#/components/schemas/userId"
              required:
                - user_id
...
  parameters:
...
    chatId:
      name: chatId
      in: path
      schema:
        $ref: "./schemas.yml#/components/schemas/chatId"
      required: true
      description: "chat id"
...
  requestBodies:
...
    chatBody:
      description: chat body
      required: true
      content:
        application/json:
          schema:
            $ref: "./schemas.yml#/components/schemas/chatInput"
    joinChatBody:
      description: join chat body
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              join_code:
                type: string
            required:
              - join_code
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT